refactor(cvtservice): extract producer_validation.go from god-file
Extract ValidateProducerResponse + validateProducerRequest into a
dedicated file. Pure file movement, no behavior change.

Part 1 of 4 in the validator_service.go split (issue #107). Blame on
moved methods survives via `git log --follow -M -C`. Methods that stay
in validator_service.go retain native blame.

Per CEO+Eng plan rev 5 at
~/.gstack/projects/sahina-cvt/ceo-plans/2026-04-18-plugin-followups.md
